Renamed some classes
